refactor plugin classes into modules
