+++ marble/marblemain.py
@@ -297,1256 +297,77 @@ except Exception:
 # Neuron Plugins: ConvTranspose1D/2D/3D
 # -----------------------------
 
-class ConvTranspose1DNeuronPlugin(Conv1DNeuronPlugin):
-    def on_init(self, neuron: "Neuron") -> None:
-        # Reuse strict rule: 5 PARAM + 1 outgoing
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 5 or len(out) != 1:
-            raise ValueError(
-                f"ConvTranspose1D neuron requires exactly 5 incoming PARAM synapses and exactly 1 outgoing synapse; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "convtranspose1d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        # Gather params and data using Conv1D helpers
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 5:
-            raise ValueError("ConvTranspose1D requires 5 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src, d_src, b_src = [s.source for s in param_incs[:5]]
-        kernel = self._to_list1d(getattr(k_src, "tensor", [])) or [1.0]
-        stride = int(self._first_scalar(getattr(s_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        padding = int(max(0.0, self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)))
-        dilation = int(self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        bias = float(self._first_scalar(getattr(b_src, "tensor", 0.0), default=0.0))
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        learn_kernel = lstore.get("kernel")
-        learn_bias = lstore.get("conv_bias")
-
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            x1 = []
-            for s in data_incs:
-                x1 += self._to_list1d(getattr(s.source, "tensor", []))
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            x1 = self._to_list1d(x)
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and str(device) == "cuda":
-            try:
-                xt = torch.tensor(x1, dtype=torch.float32, device=device).view(1, 1, -1)
-                wt = (learn_kernel.to(device) if hasattr(learn_kernel, "to") else torch.tensor(kernel, dtype=torch.float32, device=device)).view(1, 1, -1)
-                bt = (learn_bias.to(device).view(-1) if hasattr(learn_bias, "to") else torch.tensor([bias], dtype=torch.float32, device=device))
-                y = torch.nn.functional.conv_transpose1d(xt, wt, bias=bt, stride=stride, padding=padding, dilation=dilation)
-                y = y.view(-1)
-                try:
-                    report("neuron", "convtranspose1d", {"in": int(xt.numel()), "out": int(y.numel()), "k": int(wt.numel()), "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        # Pure-Python fallback
-        n = len(x1)
-        klen = len(kernel)
-        out_len = (n - 1) * stride - 2 * padding + (klen - 1) * dilation + 1
-        y_list = [0.0] * max(0, out_len)
-        if hasattr(learn_kernel, "detach"):
-            try:
-                kernel = list(learn_kernel.detach().to("cpu").view(-1).tolist())
-                klen = len(kernel)
-            except Exception:
-                pass
-        if hasattr(learn_bias, "detach"):
-            try:
-                bias = float(learn_bias.detach().to("cpu").view(-1)[0].item())
-            except Exception:
-                try:
-                    bias = float(learn_bias)
-                except Exception:
-                    pass
-        for t in range(n):
-            base = t * stride
-            for i in range(klen):
-                oi = base + i * dilation - padding
-                if 0 <= oi < len(y_list):
-                    y_list[oi] += kernel[i] * x1[t]
-        y_list = [v + bias for v in y_list]
-        try:
-            report("neuron", "convtranspose1d", {"in": n, "out": len(y_list), "k": klen, "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y_list)
-        except Exception:
-            return y_list
-
-
-class ConvTranspose2DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 5 or len(out) != 1:
-            raise ValueError(
-                f"ConvTranspose2D neuron requires exactly 5 incoming PARAM synapses and exactly 1 outgoing synapse; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "convtranspose2d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 5:
-            raise ValueError("ConvTranspose2D requires 5 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src, d_src, b_src = [s.source for s in param_incs[:5]]
-        kernel_1d = self._to_list1d(getattr(k_src, "tensor", [])) or [1.0]
-        stride = int(self._first_scalar(getattr(s_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        padding = int(max(0.0, self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)))
-        dilation = int(self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        bias = float(self._first_scalar(getattr(b_src, "tensor", 0.0), default=0.0))
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        learn_kernel = lstore.get("kernel")
-        learn_bias = lstore.get("conv_bias")
-
-        # Build 2D input from DATA rows
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            rows = [self._to_list1d(getattr(s.source, "tensor", [])) for s in data_incs]
-            width = min((len(r) for r in rows if r), default=0)
-            if width <= 0:
-                x_vals = []
-                H = W = 0
-            else:
-                rows = [r[:width] for r in rows]
-                H = len(rows)
-                W = width
-                x_vals = [v for r in rows for v in r]
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            x_vals = self._to_list1d(x)
-            N = max(1, len(x_vals))
-            rh = int(math.isqrt(N))
-            if rh * rh == N:
-                H = W = rh
-            else:
-                H, W = N, 1
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and H > 0 and W > 0 and str(device) == "cuda":
-            try:
-                xt = torch.tensor(x_vals, dtype=torch.float32, device=device).view(1, 1, H, W)
-                L = max(1, len(kernel_1d))
-                r = int(math.isqrt(L))
-                if r * r == L:
-                    kh = kw = r
-                else:
-                    kh, kw = L, 1
-                base_w = (learn_kernel.to(device) if hasattr(learn_kernel, "to") else torch.tensor(kernel_1d, dtype=torch.float32, device=device))
-                wt = base_w.view(1, 1, kh, kw)
-                bt = (learn_bias.to(device).view(-1) if hasattr(learn_bias, "to") else torch.tensor([bias], dtype=torch.float32, device=device))
-                y = torch.nn.functional.conv_transpose2d(
-                    xt, wt, bias=bt, stride=(stride, stride), padding=(padding, padding), dilation=(dilation, dilation)
-                )
-                y = y.view(-1)
-                try:
-                    report("neuron", "convtranspose2d", {"inHW": [H, W], "out": int(y.numel()), "kHW": [kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        # Pure Python fallback (simplified)
-        # Compute output dims
-        if hasattr(learn_kernel, "detach"):
-            try:
-                kernel_1d = list(learn_kernel.detach().to("cpu").view(-1).tolist())
-            except Exception:
-                pass
-        if hasattr(learn_bias, "detach"):
-            try:
-                bias = float(learn_bias.detach().to("cpu").view(-1)[0].item())
-            except Exception:
-                try:
-                    bias = float(learn_bias)
-                except Exception:
-                    pass
-        L = max(1, len(kernel_1d))
-        r = int(math.isqrt(L))
-        if r * r == L:
-            kh = kw = r
-        else:
-            kh, kw = L, 1
-        out_h = (H - 1) * stride - 2 * padding + (kh - 1) * dilation + 1
-        out_w = (W - 1) * stride - 2 * padding + (kw - 1) * dilation + 1
-        y2 = [[0.0 for _ in range(max(0, out_w))] for _ in range(max(0, out_h))]
-        for iy in range(H):
-            for ix in range(W):
-                base_y = iy * stride
-                base_x = ix * stride
-                val = x_vals[iy * W + ix]
-                for ky in range(kh):
-                    for kx in range(kw):
-                        oy = base_y + ky * dilation - padding
-                        ox = base_x + kx * dilation - padding
-                        if 0 <= oy < len(y2) and 0 <= ox < len(y2[0]):
-                            y2[oy][ox] += kernel_1d[ky * kw + kx] * val
-        y_list = [v + bias for row in y2 for v in row]
-        try:
-            report("neuron", "convtranspose2d", {"inHW": [H, W], "out": len(y_list), "kHW": [kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y_list)
-        except Exception:
-            return y_list
-
-
-class ConvTranspose3DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 5 or len(out) != 1:
-            raise ValueError(
-                f"ConvTranspose3D neuron requires exactly 5 incoming PARAM synapses and exactly 1 outgoing synapse; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "convtranspose3d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 5:
-            raise ValueError("ConvTranspose3D requires 5 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src, d_src, b_src = [s.source for s in param_incs[:5]]
-        kernel_1d = self._to_list1d(getattr(k_src, "tensor", [])) or [1.0]
-        stride = int(self._first_scalar(getattr(s_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        padding = int(max(0.0, self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)))
-        dilation = int(self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0))
-        bias = float(self._first_scalar(getattr(b_src, "tensor", 0.0), default=0.0))
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        learn_kernel = lstore.get("kernel")
-        learn_bias = lstore.get("conv_bias")
-
-        # Build 3D input from DATA slices
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            slices = []
-            dims = []
-            for s in data_incs:
-                vals = self._to_list1d(getattr(s.source, "tensor", []))
-                N = max(1, len(vals))
-                r = int(math.isqrt(N))
-                if r * r == N:
-                    H = W = r
-                else:
-                    H, W = N, 1
-                dims.append((H, W))
-                slices.append(vals[: H * W])
-            Hmin = min((h for h, _ in dims), default=0)
-            Wmin = min((w for _, w in dims), default=0)
-            x_vals = []
-            D = len(slices)
-            for sl, (h, w) in zip(slices, dims):
-                for rr in range(Hmin):
-                    row = sl[rr * w:(rr + 1) * w]
-                    x_vals.extend(row[:Wmin])
-            H, W = Hmin, Wmin
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            vals = self._to_list1d(x)
-            N = max(1, len(vals))
-            r3 = round(N ** (1.0 / 3.0))
-            if r3 > 0 and (r3 * r3 * r3) == N:
-                D = H = W = int(r3)
-                x_vals = vals[: D * H * W]
-            else:
-                D, H, W = N, 1, 1
-                x_vals = vals
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and H > 0 and W > 0 and str(device) == "cuda":
-            try:
-                xt = torch.tensor(x_vals, dtype=torch.float32, device=device).view(1, 1, D, H, W)
-                L = max(1, len(kernel_1d))
-                r = round(L ** (1.0 / 3.0))
-                if r > 0 and (r * r * r) == L:
-                    kd = kh = kw = int(r)
-                else:
-                    kd, kh, kw = L, 1, 1
-                base_w = (learn_kernel.to(device) if hasattr(learn_kernel, "to") else torch.tensor(kernel_1d, dtype=torch.float32, device=device))
-                wt = base_w.view(1, 1, kd, kh, kw)
-                bt = (learn_bias.to(device).view(-1) if hasattr(learn_bias, "to") else torch.tensor([bias], dtype=torch.float32, device=device))
-                y = torch.nn.functional.conv_transpose3d(
-                    xt, wt, bias=bt, stride=(stride, stride, stride), padding=(padding, padding, padding), dilation=(dilation, dilation, dilation)
-                )
-                y = y.view(-1)
-                try:
-                    report("neuron", "convtranspose3d", {"inDHW": [D, H, W], "out": int(y.numel()), "kDHW": [kd, kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        # Pure Python fallback
-        L = max(1, len(kernel_1d))
-        r = round(L ** (1.0 / 3.0))
-        if r > 0 and (r * r * r) == L:
-            kd = kh = kw = int(r)
-        else:
-            kd, kh, kw = L, 1, 1
-        out_d = (D - 1) * stride - 2 * padding + (kd - 1) * dilation + 1
-        out_h = (H - 1) * stride - 2 * padding + (kh - 1) * dilation + 1
-        out_w = (W - 1) * stride - 2 * padding + (kw - 1) * dilation + 1
-        y = [0.0] * max(0, out_d * out_h * out_w)
-        def idx3(a, d, h, w, H_, W_):
-            return d * (H_ * W_) + h * W_ + w
-        for iz in range(D):
-            for iy in range(H):
-                for ix in range(W):
-                    base_d = iz * stride
-                    base_y = iy * stride
-                    base_x = ix * stride
-                    val = x_vals[idx3(x_vals, 0, 0, 0, 1, 1)] if False else x_vals[iz * (H * W) + iy * W + ix]
-                    for kz in range(kd):
-                        for ky in range(kh):
-                            for kx in range(kw):
-                                od = base_d + kz * dilation - padding
-                                oh = base_y + ky * dilation - padding
-                                ow = base_x + kx * dilation - padding
-                                if 0 <= od < out_d and 0 <= oh < out_h and 0 <= ow < out_w:
-                                    y[idx3(y, od, oh, ow, out_h, out_w)] += kernel_1d[(kz * kh + ky) * kw + kx] * val
-        y = [v + bias for v in y]
-        try:
-            report("neuron", "convtranspose3d", {"inDHW": [D, H, W], "out": len(y), "kDHW": [kd, kh, kw], "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y)
-        except Exception:
-            return y
-
-
 try:
-    from .plugins.conv_transpose1d import ConvTranspose1DNeuronPlugin as _CT1
-    from .plugins.conv_transpose2d import ConvTranspose2DNeuronPlugin as _CT2
-    from .plugins.conv_transpose3d import ConvTranspose3DNeuronPlugin as _CT3
-    register_neuron_type("conv_transpose1d", _CT1())
-    register_neuron_type("conv_transpose2d", _CT2())
-    register_neuron_type("conv_transpose3d", _CT3())
+    from .plugins.conv_transpose1d import ConvTranspose1DNeuronPlugin
+    from .plugins.conv_transpose2d import ConvTranspose2DNeuronPlugin
+    from .plugins.conv_transpose3d import ConvTranspose3DNeuronPlugin
+    register_neuron_type("conv_transpose1d", ConvTranspose1DNeuronPlugin())
+    register_neuron_type("conv_transpose2d", ConvTranspose2DNeuronPlugin())
+    register_neuron_type("conv_transpose3d", ConvTranspose3DNeuronPlugin())
+    __all__ += [
+        "ConvTranspose1DNeuronPlugin",
+        "ConvTranspose2DNeuronPlugin",
+        "ConvTranspose3DNeuronPlugin",
+    ]
 except Exception:
     pass
-
-__all__ += [
-    "ConvTranspose1DNeuronPlugin",
-    "ConvTranspose2DNeuronPlugin",
-    "ConvTranspose3DNeuronPlugin",
-]
-
 
 # -----------------------------
 # Neuron Plugins: MaxPool1D/2D/3D
 # -----------------------------
 
-class MaxPool1DNeuronPlugin(Conv1DNeuronPlugin):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 3 or len(out) != 1:
-            raise ValueError(
-                f"MaxPool1D neuron requires exactly 3 incoming PARAM synapses (kernel,stride,padding) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "maxpool1d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 3:
-            raise ValueError("MaxPool1D requires 3 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src = [s.source for s in param_incs[:3]]
-        # Allow learnable hyperparameters (stored as floats, later coerced to ints)
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        lk = lstore.get("kernel_size")
-        ls = lstore.get("stride")
-        lp = lstore.get("padding")
-        try:
-            ksize = int(max(1, round(float(lk.detach().to("cpu").view(-1)[0].item())))) if hasattr(lk, "detach") else int(max(1, round(base_ks)))
-        except Exception:
-            ksize = int(max(1, round(base_ks)))
-        try:
-            stride = int(max(1, round(float(ls.detach().to("cpu").view(-1)[0].item())))) if hasattr(ls, "detach") else int(max(1, round(base_st)))
-        except Exception:
-            stride = int(max(1, round(base_st)))
-        try:
-            padding = int(max(0, round(float(lp.detach().to("cpu").view(-1)[0].item())))) if hasattr(lp, "detach") else int(max(0, round(base_pd)))
-        except Exception:
-            padding = int(max(0, round(base_pd)))
-
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            x1 = []
-            for s in data_incs:
-                x1 += self._to_list1d(getattr(s.source, "tensor", []))
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            x1 = self._to_list1d(x)
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None:
-            try:
-                xt = torch.tensor(x1, dtype=torch.float32, device=device).view(1, 1, -1)
-                y = torch.nn.functional.max_pool1d(xt, kernel_size=ksize, stride=stride, padding=padding)
-                y = y.view(-1)
-                try:
-                    report("neuron", "maxpool1d", {"in": int(xt.numel()), "out": int(y.numel()), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        # Pure Python fallback using -inf as padding
-        if padding > 0:
-            x1 = ([float("-inf")] * padding) + x1 + ([float("-inf")] * padding)
-        n = len(x1)
-        out_len = 0 if n < ksize else 1 + (n - ksize) // stride
-        y_list = []
-        for t in range(out_len):
-            base = t * stride
-            y_list.append(max(x1[base: base + ksize]))
-        try:
-            report("neuron", "maxpool1d", {"in": len(x1), "out": len(y_list), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y_list)
-        except Exception:
-            return y_list
-
-
-class MaxPool2DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 3 or len(out) != 1:
-            raise ValueError(
-                f"MaxPool2D neuron requires exactly 3 incoming PARAM synapses (kernel,stride,padding) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "maxpool2d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 3:
-            raise ValueError("MaxPool2D requires 3 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src = [s.source for s in param_incs[:3]]
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        lk = lstore.get("kernel_size")
-        ls = lstore.get("stride")
-        lp = lstore.get("padding")
-        try:
-            ksize = int(max(1, round(float(lk.detach().to("cpu").view(-1)[0].item())))) if hasattr(lk, "detach") else int(max(1, round(base_ks)))
-        except Exception:
-            ksize = int(max(1, round(base_ks)))
-        try:
-            stride = int(max(1, round(float(ls.detach().to("cpu").view(-1)[0].item())))) if hasattr(ls, "detach") else int(max(1, round(base_st)))
-        except Exception:
-            stride = int(max(1, round(base_st)))
-        try:
-            padding = int(max(0, round(float(lp.detach().to("cpu").view(-1)[0].item())))) if hasattr(lp, "detach") else int(max(0, round(base_pd)))
-        except Exception:
-            padding = int(max(0, round(base_pd)))
-
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            rows = [self._to_list1d(getattr(s.source, "tensor", [])) for s in data_incs]
-            width = min((len(r) for r in rows if r), default=0)
-            if width <= 0:
-                x_vals = []
-                H = W = 0
-            else:
-                rows = [r[:width] for r in rows]
-                H = len(rows)
-                W = width
-                x_vals = [v for r in rows for v in r]
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            x_vals = self._to_list1d(x)
-            N = max(1, len(x_vals))
-            rh = int(math.isqrt(N))
-            if rh * rh == N:
-                H = W = rh
-            else:
-                H, W = N, 1
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and H > 0 and W > 0:
-            try:
-                xt = torch.tensor(x_vals, dtype=torch.float32, device=device).view(1, 1, H, W)
-                y = torch.nn.functional.max_pool2d(xt, kernel_size=(ksize, ksize), stride=(stride, stride), padding=(padding, padding))
-                y = y.view(-1)
-                try:
-                    report("neuron", "maxpool2d", {"inHW": [H, W], "out": int(y.numel()), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        if H <= 0 or W <= 0:
-            return neuron._ensure_tensor([]) if hasattr(neuron, "_ensure_tensor") else []
-        if padding > 0:
-            padded = []
-            zero_row = [float("-inf")] * (W + 2 * padding)
-            for _ in range(padding):
-                padded.append(list(zero_row))
-            for r_ in range(H):
-                row = [float("-inf")] * padding + x_vals[r_ * W:(r_ + 1) * W] + [float("-inf")] * padding
-                padded.append(row)
-            for _ in range(padding):
-                padded.append(list(zero_row))
-            H2, W2 = len(padded), len(padded[0])
-        else:
-            padded = [x_vals[r_ * W:(r_ + 1) * W] for r_ in range(H)]
-            H2, W2 = H, W
-        out_h = 0 if H2 < ksize else 1 + (H2 - ksize) // stride
-        out_w = 0 if W2 < ksize else 1 + (W2 - ksize) // stride
-        y2 = []
-        for oy in range(out_h):
-            base_y = oy * stride
-            for ox in range(out_w):
-                base_x = ox * stride
-                m = float("-inf")
-                for ky in range(ksize):
-                    for kx in range(ksize):
-                        vy = padded[base_y + ky][base_x + kx]
-                        if vy > m:
-                            m = vy
-                y2.append(m)
-        try:
-            report("neuron", "maxpool2d", {"inHW": [H, W], "out": len(y2), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y2)
-        except Exception:
-            return y2
-
-
-class MaxPool3DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 3 or len(out) != 1:
-            raise ValueError(
-                f"MaxPool3D neuron requires exactly 3 incoming PARAM synapses (kernel,stride,padding) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "maxpool3d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 3:
-            raise ValueError("MaxPool3D requires 3 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src = [s.source for s in param_incs[:3]]
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        lk = lstore.get("kernel_size")
-        ls = lstore.get("stride")
-        lp = lstore.get("padding")
-        try:
-            ksize = int(max(1, round(float(lk.detach().to("cpu").view(-1)[0].item())))) if hasattr(lk, "detach") else int(max(1, round(base_ks)))
-        except Exception:
-            ksize = int(max(1, round(base_ks)))
-        try:
-            stride = int(max(1, round(float(ls.detach().to("cpu").view(-1)[0].item())))) if hasattr(ls, "detach") else int(max(1, round(base_st)))
-        except Exception:
-            stride = int(max(1, round(base_st)))
-        try:
-            padding = int(max(0, round(float(lp.detach().to("cpu").view(-1)[0].item())))) if hasattr(lp, "detach") else int(max(0, round(base_pd)))
-        except Exception:
-            padding = int(max(0, round(base_pd)))
-
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            slices = []
-            dims = []
-            for s in data_incs:
-                vals = self._to_list1d(getattr(s.source, "tensor", []))
-                N = max(1, len(vals))
-                r = int(math.isqrt(N))
-                if r * r == N:
-                    H = W = r
-                else:
-                    H, W = N, 1
-                dims.append((H, W))
-                slices.append(vals[: H * W])
-            Hmin = min((h for h, _ in dims), default=0)
-            Wmin = min((w for _, w in dims), default=0)
-            x_vals = []
-            D = len(slices)
-            for sl, (h, w) in zip(slices, dims):
-                for rr in range(Hmin):
-                    row = sl[rr * w:(rr + 1) * w]
-                    x_vals.extend(row[:Wmin])
-            H, W = Hmin, Wmin
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            vals = self._to_list1d(x)
-            N = max(1, len(vals))
-            r3 = round(N ** (1.0 / 3.0))
-            if r3 > 0 and (r3 * r3 * r3) == N:
-                D = H = W = int(r3)
-                x_vals = vals[: D * H * W]
-            else:
-                D, H, W = N, 1, 1
-                x_vals = vals
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and H > 0 and W > 0:
-            try:
-                xt = torch.tensor(x_vals, dtype=torch.float32, device=device).view(1, 1, D, H, W)
-                y = torch.nn.functional.max_pool3d(xt, kernel_size=(ksize, ksize, ksize), stride=(stride, stride, stride), padding=(padding, padding, padding))
-                y = y.view(-1)
-                try:
-                    report("neuron", "maxpool3d", {"inDHW": [D, H, W], "out": int(y.numel()), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-
-        # Pure Python fallback with -inf padding
-        if padding > 0 and H > 0 and W > 0:
-            D2, H2, W2 = D + 2 * padding, H + 2 * padding, W + 2 * padding
-            padval = float("-inf")
-            padded = [padval] * (D2 * H2 * W2)
-            for dz in range(D):
-                for yy in range(H):
-                    for xx in range(W):
-                        pd = dz + padding
-                        py = yy + padding
-                        px = xx + padding
-                        padded[pd * (H2 * W2) + py * W2 + px] = x_vals[dz * (H * W) + yy * W + xx]
-        else:
-            padded = list(x_vals)
-            D2, H2, W2 = D, H, W
-        out_d = 0 if D2 < ksize else 1 + (D2 - ksize) // stride
-        out_h = 0 if H2 < ksize else 1 + (H2 - ksize) // stride
-        out_w = 0 if W2 < ksize else 1 + (W2 - ksize) // stride
-        y_list = []
-        for od in range(out_d):
-            base_d = od * stride
-            for oy in range(out_h):
-                base_y = oy * stride
-                for ox in range(out_w):
-                    base_x = ox * stride
-                    m = float("-inf")
-                    for kz in range(ksize):
-                        for ky in range(ksize):
-                            for kx in range(ksize):
-                                val = padded[(base_d + kz) * (H2 * W2) + (base_y + ky) * W2 + (base_x + kx)]
-                                if val > m:
-                                    m = val
-                    y_list.append(m)
-        try:
-            report("neuron", "maxpool3d", {"inDHW": [D, H, W], "out": len(y_list), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-        except Exception:
-            pass
-        try:
-            return neuron._ensure_tensor(y_list)
-        except Exception:
-            return y_list
-
-
 try:
-    from .plugins.maxpool1d import MaxPool1DNeuronPlugin as _MP1
-    from .plugins.maxpool2d import MaxPool2DNeuronPlugin as _MP2
-    from .plugins.maxpool3d import MaxPool3DNeuronPlugin as _MP3
-    register_neuron_type("maxpool1d", _MP1())
-    register_neuron_type("maxpool2d", _MP2())
-    register_neuron_type("maxpool3d", _MP3())
+    from .plugins.maxpool1d import MaxPool1DNeuronPlugin
+    from .plugins.maxpool2d import MaxPool2DNeuronPlugin
+    from .plugins.maxpool3d import MaxPool3DNeuronPlugin
+    register_neuron_type("maxpool1d", MaxPool1DNeuronPlugin())
+    register_neuron_type("maxpool2d", MaxPool2DNeuronPlugin())
+    register_neuron_type("maxpool3d", MaxPool3DNeuronPlugin())
+    __all__ += [
+        "MaxPool1DNeuronPlugin",
+        "MaxPool2DNeuronPlugin",
+        "MaxPool3DNeuronPlugin",
+    ]
 except Exception:
     pass
-
-__all__ += [
-    "MaxPool1DNeuronPlugin",
-    "MaxPool2DNeuronPlugin",
-    "MaxPool3DNeuronPlugin",
-]
-
 
 # -----------------------------
 # Neuron Plugins: Unfold2D and Fold2D
 # -----------------------------
 
-class Unfold2DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 4 or len(out) != 1:
-            raise ValueError(
-                f"Unfold2D neuron requires exactly 4 incoming PARAM synapses (kernel,stride,padding,dilation) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "unfold2d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 4:
-            raise ValueError("Unfold2D requires 4 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src, d_src = [s.source for s in param_incs[:4]]
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        base_dl = self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        lk = lstore.get("kernel_size"); ls = lstore.get("stride"); lp = lstore.get("padding"); ld = lstore.get("dilation")
-        def to_int(val, default, minv):
-            try:
-                return int(max(minv, round(float(val.detach().to("cpu").view(-1)[0].item())))) if hasattr(val, "detach") else int(max(minv, round(default)))
-            except Exception:
-                return int(max(minv, round(default)))
-        ksize = to_int(lk, base_ks, 1)
-        stride = to_int(ls, base_st, 1)
-        padding = to_int(lp, base_pd, 0)
-        dilation = to_int(ld, base_dl, 1)
-
-        # Build 2D input from DATA rows
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            rows = [self._to_list1d(getattr(s.source, "tensor", [])) for s in data_incs]
-            width = min((len(r) for r in rows if r), default=0)
-            if width <= 0:
-                x_vals = []
-                H = W = 0
-            else:
-                rows = [r[:width] for r in rows]
-                H = len(rows)
-                W = width
-                x_vals = [v for r in rows for v in r]
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            x_vals = self._to_list1d(x)
-            N = max(1, len(x_vals))
-            r = int(math.isqrt(N))
-            H = W = r if r * r == N else (N, 1)
-            if isinstance(H, tuple):
-                H, W = H
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and H > 0 and W > 0:
-            try:
-                xt = torch.tensor(x_vals, dtype=torch.float32, device=device).view(1, 1, H, W)
-                y = torch.nn.functional.unfold(xt, kernel_size=(ksize, ksize), dilation=(dilation, dilation), padding=(padding, padding), stride=(stride, stride))
-                y = y.view(-1)
-                try:
-                    report("neuron", "unfold2d", {"inHW": [H, W], "out": int(y.numel()), "k": ksize, "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-        # Pure-Python unfold
-        if H <= 0 or W <= 0:
-            return neuron._ensure_tensor([]) if hasattr(neuron, "_ensure_tensor") else []
-        # Apply zero padding
-        if padding > 0:
-            padded = []
-            zero_row = [0.0] * (W + 2 * padding)
-            for _ in range(padding):
-                padded.append(list(zero_row))
-            for r_ in range(H):
-                row = [0.0] * padding + x_vals[r_ * W:(r_ + 1) * W] + [0.0] * padding
-                padded.append(row)
-            for _ in range(padding):
-                padded.append(list(zero_row))
-            H2, W2 = len(padded), len(padded[0])
-        else:
-            padded = [x_vals[r_ * W:(r_ + 1) * W] for r_ in range(H)]
-            H2, W2 = H, W
-        out_h = 0 if H2 < ((ksize - 1) * dilation + 1) else 1 + (H2 - ((ksize - 1) * dilation + 1)) // stride
-        out_w = 0 if W2 < ((ksize - 1) * dilation + 1) else 1 + (W2 - ((ksize - 1) * dilation + 1)) // stride
-        cols: List[float] = []
-        for oy in range(out_h):
-            base_y = oy * stride
-            for ox in range(out_w):
-                base_x = ox * stride
-                for ky in range(ksize):
-                    for kx in range(ksize):
-                        iy = base_y + ky * dilation
-                        ix = base_x + kx * dilation
-                        cols.append(padded[iy][ix])
-        try:
-            report("neuron", "unfold2d", {"inHW": [H, W], "out": len(cols), "k": ksize, "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        return neuron._ensure_tensor(cols)
-
-
-class Fold2DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 6 or len(out) != 1:
-            raise ValueError(
-                f"Fold2D neuron requires exactly 6 incoming PARAM synapses (out_h,out_w,kernel,stride,padding,dilation) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "fold2d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        if len(param_incs) < 6:
-            raise ValueError("Fold2D requires 6 incoming PARAM synapses")
-        param_incs.sort(key=self._key_src)
-        oh_src, ow_src, k_src, s_src, p_src, d_src = [s.source for s in param_incs[:6]]
-        base_oh = self._first_scalar(getattr(oh_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_ow = self._first_scalar(getattr(ow_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 1.0), default=1.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        base_dl = self._first_scalar(getattr(d_src, "tensor", 1.0), default=1.0, min_val=1.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        def to_int(val, default, minv):
-            try:
-                return int(max(minv, round(float(val.detach().to("cpu").view(-1)[0].item())))) if hasattr(val, "detach") else int(max(minv, round(default)))
-            except Exception:
-                return int(max(minv, round(default)))
-        out_h = to_int(lstore.get("out_h"), base_oh, 1)
-        out_w = to_int(lstore.get("out_w"), base_ow, 1)
-        ksize = to_int(lstore.get("kernel_size"), base_ks, 1)
-        stride = to_int(lstore.get("stride"), base_st, 1)
-        padding = to_int(lstore.get("padding"), base_pd, 0)
-        dilation = to_int(lstore.get("dilation"), base_dl, 1)
-
-        # Build columns input from DATA sources (concatenate into 1D)
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        if data_incs:
-            data_incs.sort(key=self._key_src)
-            cols = []
-            for s in data_incs:
-                cols += self._to_list1d(getattr(s.source, "tensor", []))
-        else:
-            x = input_value if input_value is not None else getattr(neuron, "tensor", [])
-            cols = self._to_list1d(x)
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        # Expect length multiple of ksize*ksize
-        kk = max(1, ksize * ksize)
-        L = len(cols) // kk if len(cols) >= kk else 0
-        if torch is not None and L > 0:
-            try:
-                ct = torch.tensor(cols[: L * kk], dtype=torch.float32, device=device).view(1, 1 * kk, L)
-                y = torch.nn.functional.fold(ct, output_size=(out_h, out_w), kernel_size=(ksize, ksize), dilation=(dilation, dilation), padding=(padding, padding), stride=(stride, stride))
-                y = y.view(-1)
-                try:
-                    report("neuron", "fold2d", {"outHW": [out_h, out_w], "in_cols": L, "k": ksize, "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-        # Pure-Python simple overlap-add
-        if L <= 0:
-            return neuron._ensure_tensor([]) if hasattr(neuron, "_ensure_tensor") else []
-        out = [[0.0 for _ in range(out_w)] for _ in range(out_h)]
-        idx = 0
-        # Compute number of sliding positions from out_h/out_w, stride/dilation/padding heuristic
-        # We place windows in raster order
-        # Determine number of windows that fits given stride
-        def npos(size):
-            # minimal model: try as many as fit when starting at 0
-            cnt = 0
-            pos = 0
-            span = (ksize - 1) * dilation + 1
-            while pos + span <= size:
-                cnt += 1
-                pos += stride
-            return cnt
-        nh = npos(out_h)
-        nw = npos(out_w)
-        total = nh * nw
-        used = min(L, total)
-        for oy in range(nh):
-            base_y = oy * stride
-            for ox in range(nw):
-                if idx >= used:
-                    break
-                base_x = ox * stride
-                # one column corresponds to ksize*ksize elements
-                for ky in range(ksize):
-                    for kx in range(ksize):
-                        iy = base_y + ky * dilation
-                        ix = base_x + kx * dilation
-                        if 0 <= iy < out_h and 0 <= ix < out_w:
-                            out[iy][ix] += cols[idx * kk + ky * ksize + kx]
-                idx += 1
-        flat = [v for row in out for v in row]
-        try:
-            report("neuron", "fold2d", {"outHW": [out_h, out_w], "in_cols": used, "k": ksize, "stride": stride, "pad": padding, "dil": dilation}, "plugins")
-        except Exception:
-            pass
-        return neuron._ensure_tensor(flat)
-
-
 try:
-    from .plugins.unfold2d import Unfold2DNeuronPlugin as _UF
-    from .plugins.fold2d import Fold2DNeuronPlugin as _FD
-    register_neuron_type("unfold2d", _UF())
-    register_neuron_type("fold2d", _FD())
+    from .plugins.unfold2d import Unfold2DNeuronPlugin
+    from .plugins.fold2d import Fold2DNeuronPlugin
+    register_neuron_type("unfold2d", Unfold2DNeuronPlugin())
+    register_neuron_type("fold2d", Fold2DNeuronPlugin())
+    __all__ += [
+        "Unfold2DNeuronPlugin",
+        "Fold2DNeuronPlugin",
+    ]
 except Exception:
     pass
-
-__all__ += ["Unfold2DNeuronPlugin", "Fold2DNeuronPlugin"]
-
 
 # -----------------------------
 # Neuron Plugins: MaxUnpool1D/2D/3D
 # -----------------------------
 
-class MaxUnpool1DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 3 or len(out) != 1:
-            raise ValueError(
-                f"MaxUnpool1D neuron requires exactly 3 incoming PARAM synapses (kernel,stride,padding) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "maxunpool1d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src = [s.source for s in param_incs[:3]]
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        def to_int(val, default, minv):
-            try:
-                return int(max(minv, round(float(val.detach().to("cpu").view(-1)[0].item())))) if hasattr(val, "detach") else int(max(minv, round(default)))
-            except Exception:
-                return int(max(minv, round(default)))
-        ksize = to_int(lstore.get("kernel_size"), base_ks, 1)
-        stride = to_int(lstore.get("stride"), base_st, 1)
-        padding = to_int(lstore.get("padding"), base_pd, 0)
-
-        # DATA sources: first = values, second = indices
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        data_incs.sort(key=self._key_src)
-        vals = self._to_list1d(getattr(data_incs[0].source, "tensor", [])) if len(data_incs) >= 1 else self._to_list1d(input_value)
-        idxs = self._to_list1d(getattr(data_incs[1].source, "tensor", [])) if len(data_incs) >= 2 else []
-
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and vals and str(device) == "cuda":
-            try:
-                vt = torch.tensor(vals, dtype=torch.float32, device=device).view(1, 1, -1)
-                it = torch.tensor([int(i) for i in idxs[: len(vals)]], dtype=torch.long, device=device).view(1, 1, -1)
-                out_len = (vt.shape[-1] - 1) * stride - 2 * padding + ksize
-                y = torch.nn.functional.max_unpool1d(vt, it, kernel_size=ksize, stride=stride, padding=padding, output_size=(1, 1, int(out_len)))
-                y = y.view(-1)
-                try:
-                    report("neuron", "maxunpool1d", {"in": int(vt.numel()), "out": int(y.numel()), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-        # Fallback: return values unchanged
-        return neuron._ensure_tensor(vals)
-
-
-class MaxUnpool2DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 3 or len(out) != 1:
-            raise ValueError(
-                f"MaxUnpool2D neuron requires exactly 3 incoming PARAM synapses (kernel,stride,padding) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "maxunpool2d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src = [s.source for s in param_incs[:3]]
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        def to_int(val, default, minv):
-            try:
-                return int(max(minv, round(float(val.detach().to("cpu").view(-1)[0].item())))) if hasattr(val, "detach") else int(max(minv, round(default)))
-            except Exception:
-                return int(max(minv, round(default)))
-        ksize = to_int(lstore.get("kernel_size"), base_ks, 1)
-        stride = to_int(lstore.get("stride"), base_st, 1)
-        padding = to_int(lstore.get("padding"), base_pd, 0)
-
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        data_incs.sort(key=self._key_src)
-        vals = []
-        idxs = []
-        if len(data_incs) >= 1:
-            vals = self._to_list1d(getattr(data_incs[0].source, "tensor", []))
-        if len(data_incs) >= 2:
-            idxs = self._to_list1d(getattr(data_incs[1].source, "tensor", []))
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and vals and str(device) == "cuda":
-            try:
-                # Infer pooled H,W as square if possible
-                N = len(vals)
-                r = int(math.isqrt(N))
-                H = W = r if r * r == N else (N, 1)
-                if isinstance(H, tuple):
-                    H, W = H
-                vt = torch.tensor(vals, dtype=torch.float32, device=device).view(1, 1, H, W)
-                it = torch.tensor([int(i) for i in idxs[: len(vals)]], dtype=torch.long, device=device).view(1, 1, H, W)
-                out_h = (H - 1) * stride - 2 * padding + ksize
-                out_w = (W - 1) * stride - 2 * padding + ksize
-                y = torch.nn.functional.max_unpool2d(vt, it, kernel_size=(ksize, ksize), stride=(stride, stride), padding=(padding, padding), output_size=(1, 1, int(out_h), int(out_w)))
-                y = y.view(-1)
-                try:
-                    report("neuron", "maxunpool2d", {"inHW": [H, W], "out": int(y.numel()), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-        return neuron._ensure_tensor(vals)
-
-
-class MaxUnpool3DNeuronPlugin(_ConvNDCommon):
-    def on_init(self, neuron: "Neuron") -> None:
-        inc = list(getattr(neuron, "incoming", []) or [])
-        out = list(getattr(neuron, "outgoing", []) or [])
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in inc if is_param(s)]
-        if len(param_incs) != 3 or len(out) != 1:
-            raise ValueError(
-                f"MaxUnpool3D neuron requires exactly 3 incoming PARAM synapses (kernel,stride,padding) and exactly 1 outgoing; got params={len(param_incs)} out={len(out)}"
-            )
-        try:
-            report("neuron", "maxunpool3d_init", {"incoming_params": len(param_incs), "outgoing": len(out)}, "plugins")
-        except Exception:
-            pass
-
-    def forward(self, neuron: "Neuron", input_value=None):
-        incoming = list(getattr(neuron, "incoming", []))
-        def is_param(s):
-            t = getattr(s, "type_name", None)
-            return isinstance(t, str) and t.startswith("param")
-        param_incs = [s for s in incoming if is_param(s)]
-        param_incs.sort(key=self._key_src)
-        k_src, s_src, p_src = [s.source for s in param_incs[:3]]
-        base_ks = self._first_scalar(getattr(k_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_st = self._first_scalar(getattr(s_src, "tensor", 2.0), default=2.0, min_val=1.0)
-        base_pd = self._first_scalar(getattr(p_src, "tensor", 0.0), default=0.0)
-        lstore = getattr(neuron, "_plugin_state", {}).get("learnable_params", {})
-        def to_int(val, default, minv):
-            try:
-                return int(max(minv, round(float(val.detach().to("cpu").view(-1)[0].item())))) if hasattr(val, "detach") else int(max(minv, round(default)))
-            except Exception:
-                return int(max(minv, round(default)))
-        ksize = to_int(lstore.get("kernel_size"), base_ks, 1)
-        stride = to_int(lstore.get("stride"), base_st, 1)
-        padding = to_int(lstore.get("padding"), base_pd, 0)
-
-        data_incs = [s for s in incoming if getattr(s, "type_name", None) == "data"]
-        data_incs.sort(key=self._key_src)
-        vals = []
-        idxs = []
-        if len(data_incs) >= 1:
-            vals = self._to_list1d(getattr(data_incs[0].source, "tensor", []))
-        if len(data_incs) >= 2:
-            idxs = self._to_list1d(getattr(data_incs[1].source, "tensor", []))
-        torch = getattr(neuron, "_torch", None)
-        device = getattr(neuron, "_device", "cpu")
-        if torch is not None and vals:
-            try:
-                N = len(vals)
-                r = int(round(N ** (1.0 / 3.0)))
-                if r > 0 and r * r * r == N:
-                    D = H = W = r
-                else:
-                    D, H, W = N, 1, 1
-                vt = torch.tensor(vals, dtype=torch.float32, device=device).view(1, 1, D, H, W)
-                it = torch.tensor([int(i) for i in idxs[: len(vals)]], dtype=torch.long, device=device).view(1, 1, D, H, W)
-                out_d = (D - 1) * stride - 2 * padding + ksize
-                out_h = (H - 1) * stride - 2 * padding + ksize
-                out_w = (W - 1) * stride - 2 * padding + ksize
-                y = torch.nn.functional.max_unpool3d(
-                    vt, it, kernel_size=(ksize, ksize, ksize), stride=(stride, stride, stride), padding=(padding, padding, padding),
-                    output_size=(1, 1, int(out_d), int(out_h), int(out_w))
-                )
-                y = y.view(-1)
-                try:
-                    report("neuron", "maxunpool3d", {"inDHW": [D, H, W], "out": int(y.numel()), "k": ksize, "stride": stride, "pad": padding}, "plugins")
-                except Exception:
-                    pass
-                return y
-            except Exception:
-                pass
-        return neuron._ensure_tensor(vals)
-
-
 try:
-    from .plugins.maxunpool1d import MaxUnpool1DNeuronPlugin as _MUP1
-    from .plugins.maxunpool2d import MaxUnpool2DNeuronPlugin as _MUP2
-    from .plugins.maxunpool3d import MaxUnpool3DNeuronPlugin as _MUP3
-    register_neuron_type("maxunpool1d", _MUP1())
-    register_neuron_type("maxunpool2d", _MUP2())
-    register_neuron_type("maxunpool3d", _MUP3())
+    from .plugins.maxunpool1d import MaxUnpool1DNeuronPlugin
+    from .plugins.maxunpool2d import MaxUnpool2DNeuronPlugin
+    from .plugins.maxunpool3d import MaxUnpool3DNeuronPlugin
+    register_neuron_type("maxunpool1d", MaxUnpool1DNeuronPlugin())
+    register_neuron_type("maxunpool2d", MaxUnpool2DNeuronPlugin())
+    register_neuron_type("maxunpool3d", MaxUnpool3DNeuronPlugin())
+    __all__ += [
+        "MaxUnpool1DNeuronPlugin",
+        "MaxUnpool2DNeuronPlugin",
+        "MaxUnpool3DNeuronPlugin",
+    ]
 except Exception:
     pass
 
-__all__ += [
-    "MaxUnpool1DNeuronPlugin",
-    "MaxUnpool2DNeuronPlugin",
-    "MaxUnpool3DNeuronPlugin",
-]
-
 # -----------------------------
-# Learning Paradigm Plugins (Brain-level orchestration)
+# Learning Paradigm Plugins (Brain-level orchestration) (Brain-level orchestration)
 # -----------------------------
 
 _PARADIGM_TYPES: Dict[str, Any] = {}
@@ -1720,223 +541,26 @@ from .plugins.synapse_noisy import NoisySynapsePlugin as _NoisySynapsePlugin  # 
 
 
 # -----------------------------
-# Wanderer Plugins: L2 Weight Penalty
+# Wanderer Plugins: L2 Weight Penalty and Others
 # -----------------------------
 
-class L2WeightPenaltyPlugin:
-    """Adds L2 penalty over visited neurons' weights and biases to the loss.
-
-    Reads lambda from wanderer._neuro_cfg['l2_lambda'] (default 0.0).
-    """
-
-    def loss(self, wanderer: "Wanderer", outputs: List[Any]):  # noqa: D401
-        torch = getattr(wanderer, "_torch", None)
-        lam = float(getattr(wanderer, "_neuro_cfg", {}).get("l2_lambda", 0.0))
-        if lam <= 0.0 or torch is None:
-            # No contribution
-            return 0.0 if torch is None else torch.tensor(0.0, device=getattr(wanderer, "_device", "cpu"))
-        total = None
-        for n in getattr(wanderer, "_visited", []) or []:
-            try:
-                w_param, b_param = wanderer._param_map[id(n)]
-                term = (w_param.view(-1) ** 2).sum() + (b_param.view(-1) ** 2).sum()
-                total = term if total is None else (total + term)
-            except Exception:
-                continue
-        if total is None:
-            return torch.tensor(0.0, device=getattr(wanderer, "_device", "cpu"))
-        return lam * total
-
-
 try:
+    from .plugins.wanderer_l2_penalty import L2WeightPenaltyPlugin
+    from .plugins.wanderer_contrastive_infonce import ContrastiveInfoNCEPlugin
+    from .plugins.wanderer_td_qlearning import TDQLearningPlugin
+    from .plugins.wanderer_distillation import DistillationPlugin
     register_wanderer_type("l2_weight_penalty", L2WeightPenaltyPlugin())
-    __all__ += ["L2WeightPenaltyPlugin"]
-except Exception:
-    pass
-
-
-# -----------------------------
-# Wanderer Plugins: Contrastive, TD Q-Learning, Distillation
-# -----------------------------
-
-class ContrastiveInfoNCEPlugin:
-    """Adds an InfoNCE-style contrastive loss across walk outputs.
-
-    Config via wanderer._neuro_cfg:
-      - contrastive_tau (temperature, default 0.1)
-      - contrastive_lambda (weight, default 1.0)
-    Positives are adjacent outputs in the same walk; negatives are all other outputs.
-    """
-
-    def _normalize(self, torch, x):
-        x = x.view(-1)
-        n = x.norm(p=2) + 1e-8
-        return x / n
-
-    def loss(self, wanderer: "Wanderer", outputs: List[Any]):  # noqa: D401
-        torch = getattr(wanderer, "_torch", None)
-        if torch is None or len(outputs) < 2:
-            return 0.0 if torch is None else torch.tensor(0.0, device=getattr(wanderer, "_device", "cpu"))
-        tau = float(getattr(wanderer, "_neuro_cfg", {}).get("contrastive_tau", 0.1))
-        w = float(getattr(wanderer, "_neuro_cfg", {}).get("contrastive_lambda", 1.0))
-        dev = getattr(wanderer, "_device", "cpu")
-        # Build matrix of normalized embeddings
-        vecs = []
-        for y in outputs:
-            if hasattr(y, "detach"):
-                v = y.detach().to(dev).float().view(-1)
-            else:
-                v = torch.tensor([float(vv) for vv in (y if isinstance(y, (list, tuple)) else [y])], dtype=torch.float32, device=dev)
-            vecs.append(self._normalize(torch, v))
-        X = torch.stack(vecs, dim=0)  # [T, D]
-        # Similarities
-        S = X @ X.t()  # [T, T]
-        S = S / max(1e-8, float(tau))
-        # For each i>=1, positive is (i, i-1)
-        loss_terms = []
-        for i in range(1, X.shape[0]):
-            logits = S[i]
-            pos = logits[i - 1]
-            # Exclude self from denominator by masking
-            mask = torch.ones_like(logits, dtype=torch.bool)
-            mask[i] = False
-            denom = torch.logsumexp(logits[mask], dim=0)
-            loss_i = - (pos - denom)
-            loss_terms.append(loss_i)
-        if not loss_terms:
-            return torch.tensor(0.0, device=dev)
-        return w * (torch.stack(loss_terms).mean())
-
-
-class TDQLearningPlugin:
-    """Tabular TD(0) Q-learning over synapses stored in synapse._plugin_state['q'].
-
-    - choose_next: epsilon-greedy over Q(synapse) for available choices.
-    - on_step: TD update for the previous chosen synapse using reward = -current per-step loss.
-    Config via wanderer._neuro_cfg: rl_epsilon (0.1), rl_alpha (0.1), rl_gamma (0.9)
-    """
-
-    def __init__(self) -> None:
-        self._last_syn: Optional["Synapse"] = None
-
-    def _q(self, syn: "Synapse") -> float:
-        st = getattr(syn, "_plugin_state", None)
-        if st is None:
-            syn._plugin_state = {}
-            st = syn._plugin_state
-        q = st.get("q", 0.0)
-        try:
-            return float(q)
-        except Exception:
-            return 0.0
-
-    def _set_q(self, syn: "Synapse", q: float) -> None:
-        st = getattr(syn, "_plugin_state", None)
-        if st is None:
-            syn._plugin_state = {}
-            st = syn._plugin_state
-        st["q"] = float(q)
-
-    def choose_next(self, wanderer: "Wanderer", current: "Neuron", choices: List[Tuple["Synapse", str]]):
-        import random as _r
-        if not choices:
-            return None, "forward"
-        eps = float(getattr(wanderer, "_neuro_cfg", {}).get("rl_epsilon", 0.1))
-        if _r.random() < eps:
-            return choices[_r.randrange(len(choices))]
-        # Greedy by Q
-        best = choices[0]
-        best_q = self._q(best[0])
-        for s, d in choices[1:]:
-            q = self._q(s)
-            if q > best_q:
-                best = (s, d); best_q = q
-        # Track chosen for TD update on next step
-        self._last_syn = best[0]
-        return best
-
-    def on_step(self, wanderer: "Wanderer", current: "Neuron", next_syn: Optional["Synapse"], direction: str, step_index: int, out_value: Any) -> None:
-        # TD update for previously chosen synapse using reward from current step loss
-        if self._last_syn is None:
-            return
-        try:
-            cur_loss_t = wanderer._walk_ctx.get("cur_loss_tensor")  # type: ignore[attr-defined]
-            r = -float(cur_loss_t.detach().to("cpu").item()) if cur_loss_t is not None else 0.0
-        except Exception:
-            r = 0.0
-        alpha = float(getattr(wanderer, "_neuro_cfg", {}).get("rl_alpha", 0.1))
-        gamma = float(getattr(wanderer, "_neuro_cfg", {}).get("rl_gamma", 0.9))
-        # Estimate max_a' Q(s',a') from the NEXT state's outgoing choices
-        max_next = 0.0
-        try:
-            if next_syn is not None:
-                next_node = next_syn.target if direction == "forward" else next_syn.source
-            else:
-                next_node = current
-            choices = wanderer._gather_choices(next_node)
-            if choices:
-                max_next = max(self._q(s) for s, _ in choices)
-        except Exception:
-            pass
-        q = self._q(self._last_syn)
-        td_target = r + gamma * max_next
-        new_q = q + alpha * (td_target - q)
-        self._set_q(self._last_syn, new_q)
-        try:
-            report("wanderer", "td_q_update", {"q": new_q, "r": r}, "plugins")
-        except Exception:
-            pass
-        # Reset last_syn only after applying update once; keep it tied to previous transition
-        self._last_syn = next_syn
-
-
-class DistillationPlugin:
-    """Adds a simple distillation loss to match a moving-average 'teacher' of outputs.
-
-    Config via wanderer._neuro_cfg:
-      - distill_lambda (default 0.1)
-      - teacher_momentum (EMA, default 0.9)
-    """
-
-    def __init__(self) -> None:
-        self._ema: Optional[Any] = None
-
-    def loss(self, wanderer: "Wanderer", outputs: List[Any]):  # noqa: D401
-        torch = getattr(wanderer, "_torch", None)
-        if torch is None or not outputs:
-            return 0.0 if torch is None else torch.tensor(0.0, device=getattr(wanderer, "_device", "cpu"))
-        dev = getattr(wanderer, "_device", "cpu")
-        lam = float(getattr(wanderer, "_neuro_cfg", {}).get("distill_lambda", 0.1))
-        mom = float(getattr(wanderer, "_neuro_cfg", {}).get("teacher_momentum", 0.9))
-        # Use the last output for current prediction
-        y = outputs[-1]
-        if not hasattr(y, "detach"):
-            y = torch.tensor([float(v) for v in (y if isinstance(y, (list, tuple)) else [y])], dtype=torch.float32, device=dev)
-        yt = y.detach().to(dev).float().view(-1)
-        if self._ema is None:
-            self._ema = yt.detach()
-        else:
-            self._ema = (mom * self._ema) + ((1.0 - mom) * yt.detach())
-        # MSE between current y and teacher ema
-        if self._ema.shape != yt.shape:
-            # Align to min length
-            m = min(int(self._ema.numel()), int(yt.numel()))
-            ema = self._ema.view(-1)[:m]
-            yv = yt.view(-1)[:m]
-        else:
-            ema = self._ema
-            yv = yt
-        return lam * ((yv - ema).pow(2).mean())
-
-
-try:
     register_wanderer_type("contrastive_infonce", ContrastiveInfoNCEPlugin())
     register_wanderer_type("td_qlearning", TDQLearningPlugin())
     register_wanderer_type("distillation", DistillationPlugin())
-    __all__ += ["ContrastiveInfoNCEPlugin", "TDQLearningPlugin", "DistillationPlugin"]
+    __all__ += [
+        "L2WeightPenaltyPlugin",
+        "ContrastiveInfoNCEPlugin",
+        "TDQLearningPlugin",
+        "DistillationPlugin",
+    ]
 except Exception:
     pass
-
 
 # -----------------------------
 # SelfAttention Routine: Adaptive Grad Clip
@@ -1999,80 +623,13 @@ except Exception:
 # Brain Training Plugin: Warmup-Decay
 # -----------------------------
 
-_BRAIN_TRAIN_TYPES: Dict[str, Any] = globals().get("_BRAIN_TRAIN_TYPES", {})  # type: ignore[var-annotated]
-
-
-def register_brain_train_type(name: str, plugin: Any) -> None:  # re-affirm in case not in scope here
-    if not isinstance(name, str) or not name:
-        raise ValueError("Brain train type name must be a non-empty string")
-    mod = getattr(getattr(plugin, "__class__", object), "__module__", "")
-    if isinstance(mod, str) and mod.startswith("marble.") and not mod.startswith("marble.plugins."):
-        raise ValueError(f"Brain training plugin '{name}' must be in marble.plugins.*; got module '{mod}'")
-    _BRAIN_TRAIN_TYPES[name] = plugin
-
-
-class WarmupDecayTrainPlugin:
-    """Per-walk scheduler that warms up LR then decays it, and grows steps.
-
-    Config (in `Brain.train(..., type_name='warmup_decay', ...)` or stacked):
-      - warmup_walks (int, default 3)
-      - base_lr (float, default 1e-2)
-      - peak_lr (float, default 5e-2)
-      - decay (float, default 0.9) multiplicative each walk after warmup
-      - start_steps (int, default 2) and step_increment (int, default 1)
-    """
-
-    def __init__(self, config: Optional[Dict[str, Any]] = None) -> None:
-        cfg = dict(config or {})
-        self.warmup = int(cfg.get("warmup_walks", 3))
-        self.base_lr = float(cfg.get("base_lr", 1e-2))
-        self.peak_lr = float(cfg.get("peak_lr", 5e-2))
-        self.decay = float(cfg.get("decay", 0.9))
-        self.start_steps = int(cfg.get("start_steps", 2))
-        self.step_inc = int(cfg.get("step_increment", 1))
-
-    def on_init(self, brain: "Brain", wanderer: "Wanderer", config: Dict[str, Any]) -> None:
-        try:
-            report("training", "warmup_decay_init", {"warmup": self.warmup}, "brain")
-        except Exception:
-            pass
-
-    def choose_start(self, brain: "Brain", wanderer: "Wanderer", i: int):
-        return None
-
-    def before_walk(self, brain: "Brain", wanderer: "Wanderer", i: int) -> Dict[str, Any]:
-        steps = self.start_steps + i * self.step_inc
-        if i < self.warmup:
-            # Linear warmup
-            t = (i + 1) / float(max(1, self.warmup))
-            lr = self.base_lr + t * (self.peak_lr - self.base_lr)
-        else:
-            # Exponential decay from peak
-            k = i - self.warmup
-            lr = self.peak_lr * (self.decay ** k)
-        try:
-            report("training", "warmup_decay_before", {"walk": i, "steps": steps, "lr": lr}, "brain")
-        except Exception:
-            pass
-        return {"max_steps": int(max(1, steps)), "lr": float(lr)}
-
-    def after_walk(self, brain: "Brain", wanderer: "Wanderer", i: int, stats: Dict[str, Any]) -> None:
-        try:
-            report("training", "warmup_decay_after", {"walk": i, "loss": stats.get("loss")}, "brain")
-        except Exception:
-            pass
-
-    def on_end(self, brain: "Brain", wanderer: "Wanderer", history: List[Dict[str, Any]]) -> Dict[str, Any]:
-        return {"warmup_decay": {"walks": len(history), "final_loss": history[-1].get("loss") if history else None}}
-
-
 try:
+    from .plugins.brain_train_warmup_decay import WarmupDecayTrainPlugin
     register_brain_train_type("warmup_decay", WarmupDecayTrainPlugin())
     __all__ += ["WarmupDecayTrainPlugin"]
 except Exception:
     pass
 
-# Example growth paradigm: grow a neuron when stuck (neuroplasticity-like behavior)
 class GrowthParadigm:
     def __init__(self, config: Optional[Dict[str, Any]] = None) -> None:
         cfg = dict(config or {})
@@ -3045,33 +1602,11 @@ def register_brain_train_type(name: str, plugin: Any) -> None:
     _BRAIN_TRAIN_TYPES[name] = plugin
 
 
-class CurriculumTrainPlugin:
-    """Brain-train plugin that increases walk max_steps across walks.
-
-    Config keys (read from the `config` dict passed to on_init):
-    - start_steps (default 1)
-    - step_increment (default 1)
-    - max_cap (optional int): if provided, cap max_steps to this value.
-    """
-    def on_init(self, brain: "Brain", wanderer: "Wanderer", config: Dict[str, Any]) -> None:
-        cfg = dict(config or {})
-        self.start_steps = int(cfg.get("start_steps", 1))
-        self.inc = int(cfg.get("step_increment", 1))
-        self.cap = cfg.get("max_cap")
-
-    def before_walk(self, brain: "Brain", wanderer: "Wanderer", i: int) -> Dict[str, Any]:
-        ms = self.start_steps + i * self.inc
-        if self.cap is not None:
-            try:
-                ms = min(ms, int(self.cap))
-            except Exception:
-                pass
-        return {"max_steps": int(ms)}
-
-
 try:
+    from .plugins.brain_train_curriculum import CurriculumTrainPlugin
     register_brain_train_type("curriculum", CurriculumTrainPlugin())
 except Exception:
+    pass
     pass
 
 
@@ -3247,915 +1782,38 @@ __all__ += ["register_brain_train_type"]
 from .wanderer import register_wanderer_type  # re-exported below
 from .wanderer import WANDERER_TYPES_REGISTRY as _WANDERER_TYPES
 try:
-    # Ensure built-in Wanderer plugins are loaded (self-register on import)
+    # Ensure built-in Wanderer plugins that self-register are loaded on import
     from .plugins.wanderer_weights import WanderAlongSynapseWeightsPlugin as _WPL  # noqa: F401
-    from .plugins.wanderer_bestpath import BestLossPathPlugin as _BPL  # noqa: F401
-    from .plugins.wanderer_altpaths import AlternatePathsCreatorPlugin as _APL  # noqa: F401
     from .plugins.wanderer_epsgreedy import EpsilonGreedyChooserPlugin as _EPL  # noqa: F401
 except Exception:
     pass
 
-
-class BestLossPathPlugin:
-    """On first choose_next in a walk, search paths up to max_steps and upweight best path synapses.
-
-    The synapse weights are increased along the best path so that a weights-driven
-    plugin (e.g., WanderAlongSynapseWeightsPlugin) will prefer it.
-    """
-    def _simulate(self, wanderer: "Wanderer", start: "Neuron", max_steps: int):
-        # DFS over choices; returns (best_loss, best_path_edges)
-        best = (float("inf"), [])
-        seen = set()
-
-        def rec(node: "Neuron", carried, depth: int, outputs: List[Any], edges: List[Tuple["Synapse", str]]):
-            nonlocal best
-            if depth >= max_steps:
-                # Evaluate
-                loss_t = wanderer._compute_loss(outputs)
-                try:
-                    loss_v = float(loss_t.detach().to("cpu").item())
-                except Exception:
-                    loss_v = float("inf")
-                if loss_v < best[0]:
-                    best = (loss_v, list(edges))
-                return
-            # Compute current output
-            out = node.forward(carried)
-            outputs2 = outputs + [out]
-            # Choices from node
-            choices = wanderer._gather_choices(node)
-            if not choices:
-                loss_t = wanderer._compute_loss(outputs2)
-                try:
-                    loss_v = float(loss_t.detach().to("cpu").item())
-                except Exception:
-                    loss_v = float("inf")
-                if loss_v < best[0]:
-                    best = (loss_v, list(edges))
-                return
-            for syn, dir_str in choices:
-                # Determine next node
-                nxt = syn.target if dir_str == "forward" else syn.source
-                rec(nxt, out, depth + 1, outputs2, edges + [(syn, dir_str)])
-
-        rec(start, None, 0, [], [])
-        return best
-
-    def _bump_weights(self, path_edges: List[Tuple["Synapse", str]], factor: float = 1.0, add: float = 1.0):
-        for syn, _ in path_edges:
-            try:
-                w = float(getattr(syn, "weight", 1.0))
-                syn.weight = w * factor + add
-            except Exception:
-                pass
-
-    def choose_next(self, wanderer: "Wanderer", current: "Neuron", choices: List[Tuple["Synapse", str]]):
-        # Run once per walk
-        if not getattr(wanderer, "_plugin_state", None):
-            wanderer._plugin_state = {}
-        if not wanderer._plugin_state.get("bestlosspath_applied"):
-            # Read config from wanderer._neuro_cfg
-            cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-            max_steps_cfg = int(cfg.get("bestlosspath_search_steps", 3))
-            bump_factor = float(cfg.get("bestlosspath_bump_factor", 1.0))
-            bump_add = float(cfg.get("bestlosspath_bump_add", 1.0))
-            best_loss, best_path = self._simulate(wanderer, current, max_steps_cfg)
-            self._bump_weights(best_path, factor=bump_factor, add=bump_add)
-            wanderer._plugin_state["bestlosspath_applied"] = True
-            try:
-                # Log node positions along path for audit
-                nodes = []
-                for syn, d in best_path:
-                    try:
-                        nodes.append({
-                            "src": getattr(getattr(syn, "source", None), "position", None),
-                            "dst": getattr(getattr(syn, "target", None), "position", None),
-                            "dir": d,
-                        })
-                    except Exception:
-                        pass
-                report("wanderer", "bestlosspath", {"best_loss": best_loss, "edges": len(best_path), "nodes": nodes}, "events")
-            except Exception:
-                pass
-        # Return current best by weight; a weights plugin stacked later will override anyway
-        if choices:
-            best = max(choices, key=lambda cd: float(getattr(cd[0], "weight", 1.0)))
-            return best
-        return (None, "forward")
-
-
 try:
+    from .plugins.wanderer_bestpath import BestLossPathPlugin
+    from .plugins.wanderer_altpaths import AlternatePathsCreatorPlugin
     register_wanderer_type("bestlosspath", BestLossPathPlugin())
-except Exception:
-    pass
-
-
-class AlternatePathsCreatorPlugin:
-    """At the end of each walk, create a new alternate path of random length and connect it to a random visited neuron.
-
-    Config keys read from `wanderer._neuro_cfg`:
-    - `altpaths_min_len` (default 2)
-    - `altpaths_max_len` (default 4)
-    """
-
-    def on_walk_end(self, wanderer: "Wanderer", stats: Dict[str, Any]) -> None:
-        import random as _rand
-        brain = wanderer.brain
-        visited = getattr(wanderer, "_visited", []) or []
-        if not visited:
-            return
-        cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-        min_len = int(cfg.get("altpaths_min_len", 2))
-        max_len = int(cfg.get("altpaths_max_len", 4))
-        if max_len < min_len:
-            max_len = min_len
-        # Per-walk limit on number of new paths
-        max_paths = int(cfg.get("altpaths_max_paths_per_walk", 1))
-        created = int(getattr(wanderer, "_plugin_state", {}).get("altpaths_created", 0))
-        if created >= max_paths:
-            return
-        length = _rand.randint(min_len, max_len)
-        # Pick an anchor neuron from the visited path
-        anchor = _rand.choice(visited)
-        # Build a chain of new neurons
-        new_nodes = []
-        try:
-            for _ in range(length):
-                # pick first free index
-                idx = None
-                for cand in brain.available_indices():
-                    try:
-                        if brain.get_neuron(cand) is None:
-                            idx = cand
-                            break
-                    except Exception:
-                        continue
-                if idx is None:
-                    # try neighbor of last node (or anchor)
-                    base = new_nodes[-1] if new_nodes else anchor
-                    bpos = getattr(base, "position", None)
-                    if isinstance(bpos, tuple):
-                        for dx, dy in [(1,0),(0,1),(-1,0),(0,-1),(1,1),(-1,-1)]:
-                            cand = (bpos[0]+dx, bpos[1]+dy) if len(bpos)>=2 else (bpos[0]+dx,)
-                            try:
-                                if brain.is_inside(cand) and brain.get_neuron(cand) is None:
-                                    idx = cand
-                                    break
-                            except Exception:
-                                continue
-                if idx is None:
-                    break
-                n = brain.add_neuron(idx, tensor=0.0)
-                new_nodes.append(n)
-        except Exception:
-            new_nodes = new_nodes
-        if not new_nodes:
-            return
-        # Connect chain
-        try:
-            prev = anchor
-            for node in new_nodes:
-                brain.connect(getattr(prev, "position"), getattr(node, "position"), direction="uni")
-                prev = node
-            report("training", "altpaths_create", {"anchor": getattr(anchor, "position", None), "len": len(new_nodes)}, "events")
-            # Track created count this walk
-            if not getattr(wanderer, "_plugin_state", None):
-                wanderer._plugin_state = {}
-            wanderer._plugin_state["altpaths_created"] = created + 1
-        except Exception:
-            pass
-
-
-try:
     register_wanderer_type("alternatepathscreator", AlternatePathsCreatorPlugin())
+    __all__ += ["BestLossPathPlugin", "AlternatePathsCreatorPlugin"]
 except Exception:
     pass
-
-
- 
-
-
-class HyperEvolutionPlugin:
-    """Evolutionary architecture search over plugin stacks and parameters without cloning the brain.
-
-    Modes (via `wanderer._neuro_cfg['hyper_evo_mode']`):
-    - "per_walk" (default): before the first training walk, run `hyper_evo_steps` evolution steps
-      using temporary changes and full rollback per step. After finishing, configure the best
-      architecture, deactivate this plugin, and proceed with normal training.
-    - "per_step": at the beginning of each training step, run `hyper_evo_steps` evolution steps,
-      configure the best architecture so far, then continue the step. This lets the search continue
-      during training.
-
-    Config (via `wanderer._neuro_cfg`):
-      - hyper_evo_steps (int, default 50): evolution steps per run
-      - hyper_eval_steps (int, default 3): bounded evaluation walk steps (with lr=0)
-      - hyper_evo_mode (str, default "per_walk"): "per_walk" | "per_step"
-
-    This search mutates across all registered plugin families (wanderer, neuroplasticity, paradigms)
-    and any numeric parameters present in `wanderer._neuro_cfg`. No plugins or parameters are excluded.
-    Each evolution step applies a temporary mutation, performs a short evaluation walk (lr=0) to get
-    loss and speed metrics, and then rolls back every change unless both metrics improved; accepted
-    mutations are accumulated into the current best genome. At the end of the run, the best genome is
-    applied persistently (no temporary stacks), and this plugin deactivates itself in per_walk mode.
-    """
-
-    def on_init(self, wanderer: "Wanderer") -> None:
-        state = getattr(wanderer, "_plugin_state", None)
-        if state is None:
-            wanderer._plugin_state = {}
-            state = wanderer._plugin_state
-        state.setdefault("hyper", {
-            "gen": 0,
-            "pop": [],  # legacy fields retained (not used by new evo loop)
-            "idx": 0,
-            "last_handle": None,
-            "last_paradigms": None,
-            "best": None,
-            "best_score": None,
-            "best_loss": None,
-            "best_speed": None,
-            "done": False,
-            "disabled": False,
-        })
-        if not state["hyper"]["pop"]:
-            self._init_population(wanderer)
-
-    def _init_population(self, wanderer: "Wanderer") -> None:
-        cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-        pop_size = int(cfg.get("hyper_pop", 8))
-        import random as _r
-        pop = []
-        wnames = list(_WANDERER_TYPES.keys())
-        nnames = list(_NEURO_TYPES.keys())
-        pnames = list(_PARADIGM_TYPES.keys()) if '_PARADIGM_TYPES' in globals() else []
-        for _ in range(pop_size):
-            genome = {
-                "w": _r.sample(wnames, k=min(len(wnames), _r.randint(0, min(3, len(wnames))))) if wnames else [],
-                "n": _r.sample(nnames, k=min(len(nnames), _r.randint(0, min(2, len(nnames))))) if nnames else [],
-                "p": _r.sample(pnames, k=min(len(pnames), _r.randint(0, min(2, len(pnames))))) if pnames else [],
-                "cfg": {},
-                "score": None,
-            }
-            # Ensure at least one wanderer plugin is present to make evolution impactful.
-            if not genome["w"] and wnames:
-                try:
-                    genome["w"] = [_r.choice(wnames)]
-                except Exception:
-                    pass
-            # Encourage at least one paradigm to be active for diversity
-            if not genome["p"] and pnames:
-                try:
-                    genome["p"] = [_r.choice(pnames)]
-                except Exception:
-                    pass
-            pop.append(genome)
-        # Seed population with a strong, commonly useful combination if available
-        try:
-            if wnames and ("bestlosspath" in wnames) and ("wanderalongsynapseweights" in wnames):
-                pop[0] = {
-                    "w": ["bestlosspath", "wanderalongsynapseweights"],
-                    "n": [],
-                    "p": [],
-                    "cfg": {},
-                    "score": None,
-                }
-        except Exception:
-            pass
-        wanderer._plugin_state["hyper"]["pop"] = pop
-        wanderer._plugin_state["hyper"]["idx"] = 0
-
-    def before_walk(self, wanderer: "Wanderer", start: "Neuron") -> None:
-        # Run evolution per mode and apply the best architecture if needed
-        st = wanderer._plugin_state["hyper"]
-        if st.get("disabled"):
-            return
-        cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-        mode = str(cfg.get("hyper_evo_mode", "per_walk")).lower()
-        steps = int(cfg.get("hyper_evo_steps", 50))
-        if mode == "per_walk" and not st.get("done"):
-            self._run_evolution(wanderer, start, steps)
-            self._apply_best_persistently(wanderer)
-            st["done"] = True
-            st["disabled"] = True
-            return
-
-        # Legacy selection/apply path retained for compatibility
-        pop = st["pop"]
-        if not pop:
-            self._init_population(wanderer)
-            pop = st["pop"]
-        # Selection strategy: exploit best-so-far or explore via tournament
-        import random as _r
-        idx = 0
-        try:
-            # Identify scored genomes
-            scored = [(i, float(g.get("score_avg", g.get("score", float("inf"))))) for i, g in enumerate(pop)]
-            scored = [(i, s) for (i, s) in scored if s != float("inf")]
-            best_idx = st.get("best_idx")
-            if scored:
-                # Track current best
-                scored.sort(key=lambda t: t[1])
-                current_best_idx = scored[0][0]
-                st["best_idx"] = current_best_idx
-                # Epsilon-greedy: mostly exploit best, sometimes explore
-                eps = 0.3
-                if best_idx is not None and _r.random() > eps:
-                    idx = int(best_idx)
-                else:
-                    # Tournament among a random subset
-                    k = max(2, min(5, len(pop)))
-                    cand = _r.sample(range(len(pop)), k=k)
-                    # Rank by available scores, unknown treated as mid-rank
-                    def score_of(i):
-                        g = pop[i]
-                        s = g.get("score_avg", g.get("score"))
-                        return float(s) if s is not None else float("inf")
-                    idx = min(cand, key=score_of)
-            else:
-                # No scores yet: random selection
-                idx = _r.randrange(0, len(pop))
-        except Exception:
-            idx = 0
-        st["idx"] = idx
-        genome = pop[idx]
-        # Apply stacks
-        handle = push_temporary_plugins(wanderer, wanderer_types=genome.get("w"), neuro_types=genome.get("n"))
-        st["last_handle"] = handle
-        # Toggle paradigms: enable selected, disable others previously managed
-        prev = st.get("last_paradigms")
-        act = []
-        try:
-            # disable previous
-            if prev:
-                for nm in prev:
-                    try:
-                        wanderer.brain.enable_paradigm(nm, enabled=False)
-                    except Exception:
-                        pass
-            for nm in genome.get("p", []) or []:
-                try:
-                    wanderer.brain.enable_paradigm(nm, enabled=True)
-                    act.append(nm)
-                except Exception:
-                    pass
-            st["last_paradigms"] = act
-        except Exception:
-            pass
-        # Snapshot current neuro config to isolate genome effects
-        try:
-            st["cfg_prev"] = dict(getattr(wanderer, "_neuro_cfg", {}) or {})
-        except Exception:
-            st["cfg_prev"] = None
-        # Snapshot current LR override if any
-        try:
-            st["lr_prev"] = getattr(wanderer, "lr_override", None)
-        except Exception:
-            st["lr_prev"] = None
-        # Merge any genome cfg params into wanderer._neuro_cfg (restored on walk end)
-        try:
-            for k, v in (genome.get("cfg") or {}).items():
-                wanderer._neuro_cfg[k] = v
-        except Exception:
-            pass
-        # If genome proposes an LR override, apply it (treated like any other evolvable param)
-        try:
-            lr_prop = (genome.get("cfg") or {}).get("lr_override", None)
-            if lr_prop is not None:
-                try:
-                    wanderer.lr_override = float(lr_prop)
-                except Exception:
-                    pass
-        except Exception:
-            pass
-
-    def on_step(self, wanderer: "Wanderer", current: "Neuron", syn: "Synapse", direction: str, step_index: int, out_value: Any) -> None:
-        # Per-step mode: run micro evolution before continuing the step
-        st = wanderer._plugin_state.get("hyper", {})
-        if st.get("disabled"):
-            return
-        cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-        mode = str(cfg.get("hyper_evo_mode", "per_walk")).lower()
-        if mode != "per_step":
-            return
-        steps = int(cfg.get("hyper_evo_steps", 50))
-        self._run_evolution(wanderer, current, steps)
-        self._apply_best_persistently(wanderer)
-        return
-
-    def on_walk_end(self, wanderer: "Wanderer", stats: Dict[str, Any]) -> None:
-        # Evaluate fitness
-        st = wanderer._plugin_state.get("hyper", {})
-        pop = st.get("pop", [])
-        idx = st.get("idx", 0) % (len(pop) if pop else 1)
-        if not pop:
-            return
-        genome = pop[idx]
-        # Compute objectives (used only for legacy population mode)
-        loss = float(stats.get("loss", 0.0))
-        steps = int(stats.get("steps", 0))
-        sm = stats.get("step_metrics", []) or []
-        # per-step time
-        dts = [m.get("dt") for m in sm if m.get("dt") is not None]
-        mean_dt = (sum(dts) / len(dts)) if dts else 0.0
-        # loss decrease speed (positive good): use -avg(delta) truncated at 0
-        deltas = [float(m.get("delta", 0.0)) for m in sm]
-        if deltas:
-            avg_delta = sum(deltas) / len(deltas)
-            speed = max(0.0, -avg_delta)
-        else:
-            speed = 0.0
-        # brain size
-        try:
-            brain_size = len(getattr(wanderer.brain, "neurons", {}))
-        except Exception:
-            brain_size = 0
-        # Score: minimize loss, mean_dt, brain_size; maximize speed => use reciprocal
-        score = (loss) + (mean_dt) + (brain_size * 1e-3) + (1.0 / (1e-6 + speed) if speed > 0 else 1.0)
-        # Update instantaneous and running-average scores for the genome
-        genome["score"] = float(score)
-        try:
-            prev_avg = float(genome.get("score_avg", score))
-            trials = int(genome.get("trials", 0)) + 1
-            new_avg = (prev_avg * (trials - 1) + float(score)) / max(1, trials)
-            genome["score_avg"] = float(new_avg)
-            genome["trials"] = trials
-        except Exception:
-            genome["score_avg"] = float(score)
-            genome["trials"] = 1
-        # Restore stacks
-        try:
-            if st.get("last_handle") is not None:
-                pop_handle = st["last_handle"]
-                pop_temporary_plugins(wanderer, pop_handle)
-                st["last_handle"] = None
-        except Exception:
-            pass
-        # Restore neuro config snapshot
-        try:
-            prev = st.get("cfg_prev", None)
-            if prev is not None:
-                wanderer._neuro_cfg = dict(prev)
-                st["cfg_prev"] = None
-        except Exception:
-            pass
-        # Restore LR override
-        try:
-            if "lr_prev" in st:
-                setattr(wanderer, "lr_override", st.get("lr_prev", None))
-                st["lr_prev"] = None
-        except Exception:
-            pass
-        # Harvest observed neuro_cfg keys to expand future mutation search space
-        try:
-            kb = st.setdefault("key_bank", set())
-            for k in (getattr(wanderer, "_neuro_cfg", {}) or {}).keys():
-                try:
-                    kb.add(str(k))
-                except Exception:
-                    pass
-        except Exception:
-            pass
-
-        # Advance and evolve at end of a population cycle (count-based, independent of chosen indices)
-        st["eval_count"] = int(st.get("eval_count", 0)) + 1
-        if st["eval_count"] >= max(1, len(pop)):
-            st["eval_count"] = 0
-            self._evolve(wanderer)
-
-    def _evolve(self, wanderer: "Wanderer") -> None:
-        import random as _r
-        cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-        # Adaptive mutation rate around configured baseline
-        base_mut = float(cfg.get("hyper_mut", 0.3))
-        st = wanderer._plugin_state["hyper"]
-        cur_mut = float(st.get("mut_rate", base_mut))
-        mut_rate = max(0.05, min(0.8, cur_mut))
-        keep = int(cfg.get("hyper_keep", 2))
-        pop = st["pop"]
-        # Sort by running-average score ascending when available (else instantaneous)
-        pop.sort(key=lambda g: float(g.get("score_avg", g.get("score", float("inf")))))
-        keepers = pop[:max(1, min(keep, len(pop)))]
-        # Track best score to adapt mutation rate
-        try:
-            best_now = float(keepers[0].get("score_avg", keepers[0].get("score", float("inf")))) if keepers else float("inf")
-            last_best = float(st.get("last_best", float("inf")))
-            if best_now < last_best:
-                st["mut_rate"] = max(0.05, mut_rate * 0.9)
-            else:
-                st["mut_rate"] = min(0.8, mut_rate * 1.1)
-            st["last_best"] = best_now
-        except Exception:
-            st["mut_rate"] = mut_rate
-        # Refill
-        wnames = list(_WANDERER_TYPES.keys())
-        nnames = list(_NEURO_TYPES.keys())
-        pnames = list(_PARADIGM_TYPES.keys()) if '_PARADIGM_TYPES' in globals() else []
-        new_pop = []
-        for g in keepers:
-            new_pop.append({"w": list(g["w"]), "n": list(g["n"]), "p": list(g["p"]), "cfg": dict(g.get("cfg") or {}), "score": None, "score_avg": g.get("score_avg"), "trials": g.get("trials")})
-        while len(new_pop) < len(pop):
-            # Crossover (pick two parents)
-            a, b = _r.sample(keepers, k=2) if len(keepers) >= 2 else (keepers[0], keepers[0])
-            child = {
-                "w": list(set(_r.sample(a["w"] + b["w"], k=min(len(a["w"] + b["w"]), _r.randint(0, min(3, len(wnames))))))) if (a["w"] or b["w"]) else [],
-                "n": list(set(_r.sample(a["n"] + b["n"], k=min(len(a["n"] + b["n"]), _r.randint(0, min(2, len(nnames))))))) if (a["n"] or b["n"]) else [],
-                "p": list(set(_r.sample(a["p"] + b["p"], k=min(len(a["p"] + b["p"]), _r.randint(0, min(2, len(pnames))))))) if (a["p"] or b["p"]) else [],
-                "cfg": dict(a.get("cfg") or {}),
-                "score": None,
-            }
-            # Mutate types
-            if _r.random() < mut_rate and wnames:
-                # flip one wanderer plugin
-                choice = _r.choice(wnames)
-                if choice in child["w"]:
-                    child["w"].remove(choice)
-                else:
-                    child["w"].append(choice)
-            if _r.random() < mut_rate and nnames:
-                choice = _r.choice(nnames)
-                if choice in child["n"]:
-                    child["n"].remove(choice)
-                else:
-                    child["n"].append(choice)
-            if _r.random() < mut_rate and pnames:
-                choice = _r.choice(pnames)
-                if choice in child["p"]:
-                    child["p"].remove(choice)
-                else:
-                    child["p"].append(choice)
-        # Mutate generic numeric cfg entries over a broad key bank
-        keys = set((child.get("cfg") or {}).keys())
-        try:
-            keys |= set(getattr(wanderer, "_neuro_cfg", {}).keys())
-        except Exception:
-            pass
-        try:
-            kb = st.get("key_bank")
-            if kb:
-                keys |= set(kb)
-        except Exception:
-            pass
-        # Seed known useful numeric keys into the search space without excluding any others
-        keys.add("lr_override")
-        keys = [k for k in set(keys) if isinstance(k, str)]
-        if keys and _r.random() < mut_rate:
-            k = _r.choice(keys)
-            try:
-                base = float((child.get("cfg", {}).get(k) if k in (child.get("cfg") or {}) else getattr(wanderer, "_neuro_cfg", {}).get(k, 0.0)))
-                scale = 1.0 + ((_r.random() * 2.0 - 1.0) * (0.5 + 0.5 * (mut_rate - 0.05)))
-                child["cfg"][k] = base * scale
-            except Exception:
-                pass
-            new_pop.append(child)
-        st["pop"] = new_pop
-        st["gen"] = st.get("gen", 0) + 1
-        st["idx"] = 0
-
-    # ---------------- Evolution helpers (no cloning, full rollback) ----------------
-    def _snapshot_graph(self, brain: "Brain") -> Dict[str, Any]:
-        try:
-            neurons = set(getattr(brain, "neurons", {}).keys())
-        except Exception:
-            neurons = set()
-        try:
-            syns = list(getattr(brain, "synapses", []) or [])
-            syn_ids = set(id(s) for s in syns)
-            syn_weights = {id(s): float(getattr(s, "weight", 1.0)) for s in syns}
-        except Exception:
-            syn_ids = set()
-            syn_weights = {}
-        return {"neurons": neurons, "syn_ids": syn_ids, "syn_w": syn_weights}
-
-    def _restore_graph(self, brain: "Brain", snap: Dict[str, Any]) -> None:
-        try:
-            # Remove newly added synapses
-            current_syns = list(getattr(brain, "synapses", []) or [])
-            before_ids = snap.get("syn_ids", set())
-            for s in current_syns:
-                if id(s) not in before_ids:
-                    try:
-                        brain.remove_synapse(s)
-                    except Exception:
-                        pass
-            # Remove newly added neurons
-            before_neurons = snap.get("neurons", set())
-            for pos, n in list(getattr(brain, "neurons", {}).items()):
-                if pos not in before_neurons:
-                    try:
-                        brain.remove_neuron(n)
-                    except Exception:
-                        pass
-            # Restore original synapse weights
-            for s in list(getattr(brain, "synapses", []) or []):
-                sid = id(s)
-                if sid in snap.get("syn_w", {}):
-                    try:
-                        s.weight = float(snap["syn_w"][sid])
-                    except Exception:
-                        pass
-        except Exception:
-            pass
-
-    def _apply_genome_temp(self, wanderer: "Wanderer", genome: Dict[str, Any]) -> Dict[str, Any]:
-        # Apply stacks and paradigms temporarily; merge cfg; return a handle for rollback
-        handle = push_temporary_plugins(wanderer, wanderer_types=genome.get("w"), neuro_types=genome.get("n"))
-        prev_cfg = dict(getattr(wanderer, "_neuro_cfg", {}) or {})
-        toggled = []
-        try:
-            for nm in genome.get("p", []) or []:
-                wanderer.brain.enable_paradigm(nm, enabled=True)
-                toggled.append(nm)
-        except Exception:
-            pass
-        try:
-            for k, v in (genome.get("cfg") or {}).items():
-                wanderer._neuro_cfg[k] = v
-        except Exception:
-            pass
-        return {"handle": handle, "cfg_prev": prev_cfg, "paradigms": toggled}
-
-    def _rollback_temp(self, wanderer: "Wanderer", temp: Dict[str, Any]) -> None:
-        try:
-            if temp.get("handle") is not None:
-                pop_temporary_plugins(wanderer, temp["handle"])
-        except Exception:
-            pass
-        try:
-            for nm in temp.get("paradigms", []) or []:
-                wanderer.brain.enable_paradigm(nm, enabled=False)
-        except Exception:
-            pass
-        try:
-            wanderer._neuro_cfg = temp.get("cfg_prev", {})
-        except Exception:
-            pass
-
-    def _short_eval(self, wanderer: "Wanderer", start: "Neuron", max_steps: int) -> Dict[str, float]:
-        # Run a short walk with lr=0 to measure loss and speed proxy; rollback handled by caller
-        stats = {"loss": float("inf"), "speed": 0.0}
-        try:
-            res = wanderer.walk(max_steps=max(1, int(max_steps)), start=start, lr=0.0)
-            # Compute speed proxy from res.step_metrics
-            dts = res.get("step_metrics", []) or []
-            deltas = [float(m.get("delta", 0.0)) for m in dts]
-            speed = max(0.0, - (sum(deltas) / max(1, len(deltas)))) if deltas else 0.0
-            stats = {"loss": float(res.get("loss", 0.0)), "speed": float(speed)}
-        except Exception:
-            pass
-        return stats
-
-    def _mutate(self, genome: Dict[str, Any], wanderer: "Wanderer", rate: float) -> Dict[str, Any]:
-        import random as _r
-        g = {"w": list(genome.get("w", [])), "n": list(genome.get("n", [])), "p": list(genome.get("p", [])), "cfg": dict(genome.get("cfg", {}))}
-        wnames = list(_WANDERER_TYPES.keys())
-        nnames = list(_NEURO_TYPES.keys())
-        pnames = list(_PARADIGM_TYPES.keys()) if '_PARADIGM_TYPES' in globals() else []
-        # With rate, add/remove one plugin from each family (no specific selection bias)
-        if wnames and _r.random() < rate:
-            nm = _r.choice(wnames)
-            if nm in g["w"] and _r.random() < 0.5:
-                g["w"].remove(nm)
-            else:
-                if nm not in g["w"]:
-                    g["w"].append(nm)
-        if nnames and _r.random() < rate:
-            nm = _r.choice(nnames)
-            if nm in g["n"] and _r.random() < 0.5:
-                g["n"].remove(nm)
-            else:
-                if nm not in g["n"]:
-                    g["n"].append(nm)
-        if pnames and _r.random() < rate:
-            nm = _r.choice(pnames)
-            if nm in g["p"] and _r.random() < 0.5:
-                g["p"].remove(nm)
-            else:
-                if nm not in g["p"]:
-                    g["p"].append(nm)
-        # Numeric param tweak: pick an observed key or create a neutral one
-        keys = set(g["cfg"].keys())
-        try:
-            keys |= set((getattr(wanderer, "_neuro_cfg", {}) or {}).keys())
-        except Exception:
-            pass
-        if keys and _r.random() < rate:
-            k = _r.choice(list(keys))
-            try:
-                base = float(g["cfg"].get(k, getattr(wanderer, "_neuro_cfg", {}).get(k, 0.0)))
-                scale = 1.0 + ((_r.random() * 2.0 - 1.0) * 0.5)
-                g["cfg"][k] = base * scale
-            except Exception:
-                pass
-        return g
-
-    def _apply_best_persistently(self, wanderer: "Wanderer") -> None:
-        st = wanderer._plugin_state.get("hyper", {})
-        best = st.get("best")
-        if not best:
-            return
-        try:
-            # Replace stacks
-            wanderer._wplugins = []
-            for nm in best.get("w", []) or []:
-                plug = _WANDERER_TYPES.get(str(nm))
-                if plug is not None:
-                    wanderer._wplugins.append(plug)
-            wanderer._neuro_plugins = []
-            for nm in best.get("n", []) or []:
-                nplug = _NEURO_TYPES.get(str(nm))
-                if nplug is not None:
-                    wanderer._neuro_plugins.append(nplug)
-            # Enable paradigms
-            for nm in best.get("p", []) or []:
-                try:
-                    wanderer.brain.enable_paradigm(nm, enabled=True)
-                except Exception:
-                    pass
-            # Merge cfg
-            for k, v in (best.get("cfg") or {}).items():
-                wanderer._neuro_cfg[k] = v
-        except Exception:
-            pass
-
-    def _run_evolution(self, wanderer: "Wanderer", start: "Neuron", steps: int) -> None:
-        cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-        eval_steps = int(cfg.get("hyper_eval_steps", 3))
-        mut_rate = float(cfg.get("hyper_mut", 0.3))
-        st = wanderer._plugin_state.get("hyper", {})
-        # Initialize best genome from current stacks and cfg
-        base = {"w": [], "n": [], "p": [], "cfg": dict(cfg)}
-        # Evaluate base
-        base_snap = self._snapshot_graph(wanderer.brain)
-        temp = self._apply_genome_temp(wanderer, base)
-        base_stats = self._short_eval(wanderer, start, eval_steps)
-        self._rollback_temp(wanderer, temp)
-        self._restore_graph(wanderer.brain, base_snap)
-        best = base
-        best_loss = float(base_stats.get("loss", float("inf")))
-        best_speed = float(base_stats.get("speed", 0.0))
-        for _ in range(max(1, int(steps))):
-            cand = self._mutate(best, wanderer, rate=mut_rate)
-            snap = self._snapshot_graph(wanderer.brain)
-            temp = self._apply_genome_temp(wanderer, cand)
-            stats = self._short_eval(wanderer, start, eval_steps)
-            self._rollback_temp(wanderer, temp)
-            self._restore_graph(wanderer.brain, snap)
-            cand_loss = float(stats.get("loss", float("inf")))
-            cand_speed = float(stats.get("speed", 0.0))
-            # Accept only if BOTH metrics improve (strict dominance) per user objective
-            if cand_loss < best_loss and cand_speed > best_speed:
-                best = cand
-                best_loss = cand_loss
-                best_speed = cand_speed
-        # Persist best in plugin state
-        try:
-            st["best"] = best
-            st["best_loss"] = best_loss
-            st["best_speed"] = best_speed
-        except Exception:
-            pass
-
 
 try:
+    from .plugins.wanderer_hyperevolution import HyperEvolutionPlugin
     register_wanderer_type("hyperEvolution", HyperEvolutionPlugin())
+    __all__ += ["HyperEvolutionPlugin"]
 except Exception:
     pass
-
 
 # Neuroplasticity plugin registry (moved). Import registry and registrar.
 from .wanderer import register_neuroplasticity_type  # re-exported below
 from .wanderer import NEURO_TYPES_REGISTRY as _NEURO_TYPES
 
 
-class BaseNeuroplasticityPlugin:
-    def on_init(self, wanderer: "Wanderer") -> None:
-        try:
-            report("neuroplasticity", "init", {"type": "base"}, "events")
-        except Exception:
-            pass
-
-    def on_step(self, wanderer: "Wanderer", current: "Neuron", syn: "Synapse", direction: str, step_index: int, out_value: Any) -> None:
-        try:
-            report("neuroplasticity", "step", {"dir": direction, "step": step_index}, "events")
-        except Exception:
-            pass
-        # Optional growth during walk when stuck
-        try:
-            cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-            grow_on_step = bool(cfg.get("grow_on_step_when_stuck", False))
-            max_new = int(cfg.get("max_new_per_walk", 1))
-            if not grow_on_step:
-                return
-            if getattr(current, "outgoing", None) and len(current.outgoing) > 0:
-                return
-            cur_new = int(wanderer._plugin_state.get("neuro_new_added", 0))
-            if cur_new >= max_new:
-                return
-            brain = wanderer.brain
-            avail = []
-            try:
-                avail = brain.available_indices()
-            except Exception:
-                pass
-            if not avail:
-                return
-            last_pos = getattr(current, "position", None)
-            chosen = None
-            for cand in avail:
-                try:
-                    if cand != last_pos and brain.get_neuron(cand) is None:
-                        chosen = cand
-                        break
-                except Exception:
-                    continue
-            if chosen is None:
-                chosen = avail[0]
-            try:
-                brain.add_neuron(chosen, tensor=0.0)
-            except Exception:
-                pass
-            try:
-                brain.connect(getattr(current, "position"), chosen, direction="uni")
-            except Exception:
-                return
-            wanderer._plugin_state["neuro_new_added"] = cur_new + 1
-            report("neuroplasticity", "grow_step", {"from": getattr(current, "position", None), "to": chosen}, "events")
-        except Exception:
-            pass
-
-    def on_walk_end(self, wanderer: "Wanderer", stats: Dict[str, Any]) -> None:
-        try:
-            visited = getattr(wanderer, "_visited", [])
-            if not visited:
-                return
-            last = visited[-1]
-            cfg = getattr(wanderer, "_neuro_cfg", {}) or {}
-            grow_if_no_out = bool(cfg.get("grow_if_no_outgoing", True))
-            max_new = int(cfg.get("max_new_per_walk", 1))
-            enable_prune = bool(cfg.get("enable_prune", False))
-            prune_gt = cfg.get("prune_if_outgoing_gt") if "prune_if_outgoing_gt" in cfg else None
-            adjust_bias = cfg.get("adjust_bias_on_loss")
-
-            # Adjust bias slightly based on loss sign if requested
-            if adjust_bias is not None:
-                try:
-                    delta = -float(stats.get("loss", 0.0))
-                    last.bias = float(last.bias) + (float(adjust_bias) * (1.0 if delta < 0 else -1.0))
-                except Exception:
-                    pass
-
-            # Prune if too many outgoing
-            if enable_prune and prune_gt is not None:
-                try:
-                    if len(getattr(last, "outgoing", [])) > int(prune_gt):
-                        syn = last.outgoing[-1]
-                        wanderer.brain.remove_synapse(syn)
-                        report("neuroplasticity", "prune", {"from": getattr(last, "position", None)}, "events")
-                except Exception:
-                    pass
-
-            if getattr(last, "outgoing", None) and len(last.outgoing) > 0:
-                return
-            if not grow_if_no_out:
-                return
-            cur_new = int(wanderer._plugin_state.get("neuro_new_added", 0))
-            if cur_new >= max_new:
-                return
-            brain = wanderer.brain
-            try:
-                avail = brain.available_indices()
-            except Exception:
-                avail = []
-            if not avail:
-                return
-            # Choose an unused index different from the last neuron's position
-            last_pos = getattr(last, "position", None)
-            chosen = None
-            try:
-                for cand in avail:
-                    if cand != last_pos and brain.get_neuron(cand) is None:
-                        chosen = cand
-                        break
-            except Exception:
-                chosen = None
-            if chosen is None:
-                # Fallback to the first available even if used; handle errors gracefully
-                chosen = avail[0]
-            try:
-                brain.add_neuron(chosen, tensor=0.0)
-            except Exception:
-                # If already exists, continue by connecting only if possible
-                pass
-            idx = chosen
-            brain.connect(getattr(last, "position"), idx, direction="uni")
-            wanderer._plugin_state["neuro_new_added"] = cur_new + 1
-            report("neuroplasticity", "grow", {"from": getattr(last, "position", None), "to": idx}, "events")
-        except Exception:
-            pass
-
-
-# Register base plugin by default
-register_neuroplasticity_type("base", BaseNeuroplasticityPlugin())
+try:
+    from .plugins.neuroplasticity_base import BaseNeuroplasticityPlugin
+    __all__ += ["BaseNeuroplasticityPlugin"]
+except Exception:
+    pass
 
 from .wanderer import Wanderer
